backend/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
distlib==0.3.7
Django==4.2.3
djangorestframework==3.14.0
filelock==3.12.2
pipenv==2023.7.23
platformdirs==3.9.1
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
virtualenv==20.24.2
virtualenv-clone==0.5.7
